pyproject.toml
[project]
name = "gpt-oss-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "openai>=1.99.9",
    "python-dotenv>=1.1.1",
    "typing-extensions>=4.7",
    "uvicorn[standard]>=0.30.6",
]

[dependency-groups]
dev = [
    "basedpyright>=1.31.3",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
    "ruff>=0.12.10",
]
